make Google Artifact Registry work again
before this, the blob requests were getting redirect to the proxy itself, due
to

    location: /artifacts-download/...

header and this route is unregistered with this proxy. So it was trying to
route the request to the AR URL of the image (for the pleasure of viewing image
details from the browser on Cloud Console), and therefore corrupting the blob
download link.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>